move payments components into feature folder
