+++ src/client/pages/Payments.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from 'react';
-import PaymentList from '../resource-components/payments/PaymentList';
-import type { Eob, Deposit } from '../../../server/types/payment.types';
+import PaymentList from '../features/payments/components/PaymentList';
+import type { Eob, Deposit } from '../../server/types/payment.types';
 
 const Payments = () => {
   type Resource = 'eobs' | 'deposits' | 'matched';
